Generalize and harden ST-sync-context
