+++ src/HttpClientFactory/Http/test/Microsoft.Extensions.Http.Tests.csproj
@@ -9,4 +9,8 @@
     <Reference Include="Microsoft.Extensions.DependencyInjection" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Compile Include="..\..\..\Shared\test\SingleThreadedSynchronizationContext.cs" Link="Shared\SingleThreadedSynchronizationContext.cs"/>
+  </ItemGroup>
+
 </Project>
